+++ ui/app/src/views/RewardsPage/components/ClaimRewardsModal.tsx
@@ -16,6 +16,7 @@ import { DistributionType } from "../../../../../core/src/generated/proto/sifnod
 import { RewardsChart } from "./RewardsChart";
 import AssetIcon from "@/components/AssetIcon";
 import { flagsStore } from "@/store/modules/flags";
+import { RewardProgramParticipant } from "../useRewardsPageData";
 
 const formatRowanNumber = (n?: number) => {
   if (n == null) return "0";
@@ -37,7 +38,7 @@ export default defineComponent({
   props: {
     address: { type: String, required: true },
     userData: {
-      type: Object as PropType<CryptoeconomicsUserData>,
+      type: Object as PropType<RewardProgramParticipant>,
       required: true,
     },
     summaryAPY: { type: Number },
@@ -64,15 +65,14 @@ export default defineComponent({
             ? DistributionType.DISTRIBUTION_TYPE_LIQUIDITY_MINING
             : DistributionType.DISTRIBUTION_TYPE_VALIDATOR_SUBSIDY,
         fromAddress: props.address,
-        rewardProgramName: "COSMOS_IBC_REWARDS_V1",
+        rewardProgramName: "harvest",
       });
       if (status.state === "accepted") {
         useCore().services.bus.dispatch({
           type: "SuccessEvent",
           payload: {
             message: `Claimed ${formatRowanNumber(
-              props.userData?.user
-                ?.totalClaimableCommissionsAndClaimableRewards,
+              props.userData?.totalClaimableCommissionsAndClaimableRewards,
             )} ROWAN of rewards`,
           },
         });
@@ -84,8 +84,8 @@ export default defineComponent({
       return (
         (props.summaryAPY || 0) *
         0.01 *
-        (props.userData?.user?.yearsToMaturity || 0) *
-        (props.userData?.user?.totalDepositedAmount || 0)
+        (props.userData?.yearsToMaturity || 0) *
+        (props.userData?.totalDepositedAmount || 0)
       );
     });
 
@@ -100,8 +100,7 @@ export default defineComponent({
             "Claimable Rewards Today",
             <span class="flex items-center font-mono">
               {formatRowanNumber(
-                props.userData?.user
-                  ?.totalClaimableCommissionsAndClaimableRewards,
+                props.userData?.totalClaimableCommissionsAndClaimableRewards,
               )}
               {
                 <TokenIcon
@@ -112,12 +111,14 @@ export default defineComponent({
               }
             </span>,
           ],
-          [
-            "Maturity Date",
-            props.userData?.user?.maturityDate.toLocaleDateString() +
-              ", " +
-              props.userData?.user?.maturityDate.toLocaleTimeString(),
-          ],
+          props.userData?.maturityDateMs
+            ? [
+                "Maturity Date",
+                new Date(props.userData?.maturityDateMs).toLocaleDateString() +
+                  ", " +
+                  new Date(props.userData?.maturityDateMs).toLocaleTimeString(),
+              ]
+            : null,
           // [
           //   <span class="flex items-center">Projected Full Reward</span>,
           //   <span
@@ -215,6 +216,19 @@ export default defineComponent({
             Are you sure you want to claim your rewards?
           </p>
           <p class="mt-[10px]">
+            Claims are paid out on Friday of each week. Once your funds are
+            dispensed, you may be interested in staking or delegating your
+            earnings with{" "}
+            <a
+              class="underline font-semibold"
+              href="https://docs.sifchain.finance/resources/rewards-programs#block-rewards"
+            >
+              Sifchain's block rewards program
+            </a>
+            .
+            <br />
+          </p>
+          {/* <p class="mt-[10px]">
             If you claim your rewards now, you will accrue less rewards in total
             on your currently pooled assets than if you wait to claim until your
             Maturity Date. For more information about our rewards program,{" "}
@@ -225,9 +239,9 @@ export default defineComponent({
               class="underline"
             >
               click here
-            </a>
-            .
-            {flagsStore.state.claimsGraph && (
+            </a> */}
+          {/* . */}
+          {/* {flagsStore.state.claimsGraph && (
               <div class="mt-[32px]">
                 <RewardsChart
                   rewardsAtMaturityAfterClaim={
@@ -236,14 +250,13 @@ export default defineComponent({
                   userData={props.userData}
                 />
               </div>
-            )}
-            <Form.Details class="mt-[24px]" details={detailsRef.value} />
-          </p>
+            )} */}
+          {/* <Form.Details class="mt-[24px]" details={detailsRef.value} />
+          </p> */}
           <Button.CallToAction class="mt-[10px]" onClick={handleClaimRewards}>
             Claim{" "}
             {formatRowanNumber(
-              props.userData?.user
-                ?.totalClaimableCommissionsAndClaimableRewards,
+              props.userData?.totalClaimableCommissionsAndClaimableRewards,
             )}{" "}
             Rowan
           </Button.CallToAction>